reorganizing scripts and scenes
