+++ Assets/Prefabs/2D.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4e3686719caf37b45b92bf8f0ce9acc3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
