+++ Assets/Prefabs/3D.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f5afd14962814cd4598ece20296f3f26
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
