+++ Assets/Scenes/2D.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 52c8ba3ef97c1cf4d88bb7b414a258c6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
