+++ Assets/Scripts/2D.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: 74211ff54e5949b4e9f3489c5347d738
+guid: 5963bc5b93131b74da3510c89088dea1
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
