+++ ProjectSettings/EditorBuildSettings.asset
@@ -6,6 +6,6 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/2D Brute Force.unity
+    path: Assets/Scenes/2D/2D Brute Force.unity
     guid: 9e31b40a2caa83f4198dc37d64607eec
   m_configObjects: {}
